* Make compassview rotation value int, not float

+++ compassviewlib/src/main/java/com/gospelware/compassviewlib/OnRotationChangeListener.java
@@ -1,5 +1,5 @@
 package com.gospelware.compassviewlib;
 
 public interface OnRotationChangeListener {
-  void rotationChanged(float oldRotation, float newRotation);
+  void rotationChanged(int oldRotation, int newRotation);
 }
